feat: add lz4

+++ src/libs/bin_decoder/lv_bin_decoder.c
@@ -1156,9 +1156,10 @@ static lv_result_t decompress_image(lv_image_decoder_dsc_t * dsc, const lv_image
         int len;
         len = LZ4_decompress_safe(input, output, input_len, out_len);
         if(len < 0 || (uint32_t)len != compressed->decompressed_size) {
-            LV_LOG_WARN("Decompress failed: %" LV_PRId32 ", got: %" LV_PRId32, out_len, len);
-            lv_draw_buf_destroy(decompressed);
-            return LV_RESULT_INVALID;
+          LV_LOG_WARN("Decompress failed: %" LV_PRId32 ", got: %" LV_PRId32,
+                      out_len, (uint32_t)len);
+          lv_draw_buf_destroy(decompressed);
+          return LV_RESULT_INVALID;
         }
 #else
         LV_LOG_WARN("LZ4 decompress is not enabled");
